+++ src/lib/FlightTasks/tasks/AutoMapper2/FlightTaskAutoMapper2.cpp
@@ -95,7 +95,8 @@ bool FlightTaskAutoMapper2::update()
 
 	// during mission and reposition, raise the landing gears but only
 	// if altitude is high enough
-	if (_mission_gear == landing_gear_s::GEAR_DOWN || highEnoughForLandingGear(2.0f, _alt_above_ground)) {
+	if (_mission_gear == landing_gear_s::GEAR_DOWN || (_param_mpc_gear_auto.get()
+			&& highEnoughForLandingGear(2.0f, _alt_above_ground))) {
 		_gear.landing_gear = _mission_gear;
 	}