+++ src/lib/FlightTasks/tasks/FlightTask/FlightTask.hpp
@@ -249,6 +249,8 @@ protected:
 					(ParamFloat<px4::params::MPC_XY_VEL_MAX>) _param_mpc_xy_vel_max,
 					(ParamFloat<px4::params::MPC_Z_VEL_MAX_DN>) _param_mpc_z_vel_max_dn,
 					(ParamFloat<px4::params::MPC_Z_VEL_MAX_UP>) _param_mpc_z_vel_max_up,
-					(ParamFloat<px4::params::MPC_TILTMAX_AIR>) _param_mpc_tiltmax_air
+					(ParamFloat<px4::params::MPC_TILTMAX_AIR>) _param_mpc_tiltmax_air,
+					(ParamInt<px4::params::MPC_GEAR_AUTO>) _param_mpc_gear_auto
+
 				       )
 };